개행추가 및 리팩토링
