+++ src/js/constants/Messege.js
@@ -1,0 +1,4 @@
+const INPUT_OPERATION_WITHOUT_NUMBER_ERROR = '숫자를 먼저 입력한 후 연산자를 입력해주세요!'
+const MAX_LENGTH_NUMBERS_EXCEEDED_ERROR = '숫자는 세 자리까지만 입력 가능합니다!';
+
+export { INPUT_OPERATION_WITHOUT_NUMBER_ERROR, MAX_LENGTH_NUMBERS_EXCEEDED_ERROR };
