+++ src/js/utils/util.js
@@ -1,0 +1,3 @@
+const $ = (selector) => document.querySelector(selector);
+
+export default $;
